Refactor API and generate docs with swagger
